\name{fcd.cluster}
\alias{fcd.cluster}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Clustering the estimators along the path.
}
\description{
To use k-means to cluster the estimators along the path and get the community labels.
}
\usage{
fcd.cluster(obj, K = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{
it is an object generated by \code{\link{fcd.start}}, 
}
  \item{K}{
input integer -- the pre-specified number of communities, with the default value 2.
}
}
\value{
an array of community labels list, of dimension number of non-isolated nodes x number of effective estimators. Each entry has value from 1 to K, as an index of the community label. Notice, the community labels are usually permutation-invariant.
}
\references{
Yang Feng, Richard J. Samworth and Yi Yu, Fused Community Detection, manuscript.
}
\author{
Yang Feng, Richard J. Samworth and Yi Yu
}
